DDAD/DDAD-main/requirements.txt

kornia==0.6.12
matplotlib==3.7.1
#numpy==1.24.3
numpy
omegaconf==2.1.2
opencv-python-headless==4.5.5.64
#pandas==2.0.1
pandas
#Pillow==9.5.0
Pillow
scikit-image
scikit-learn
#scipy==1.10.1
scipy
#sklearn
#torch==2.0.1
torch
torchmetrics
torchvision
#torchmetrics==0.11.4
#torchvision==0.15.2
